requirements.txt
langchain
langchain-openai
langchain-community==0.0.38
langgraph==0.0.27
openai
chromadb==0.4.22
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
